prost-type: Impl `Hash + Eq` traits on `Timestamp`. (#509)
* [prost-type] Impl Hash for Timestamp

From what I'm reading it would fulfil the requirements for hash:
> Thankfully, you won’t need to worry about upholding this property when deriving both `Eq` and `Hash` with `#[derive(PartialEq, Eq, Hash)]`.

The edge case is when the `Timestamp` is not normalized yet and would have mutated during the `normalize()` call. However since it's not implementing `Eq` it should still be sound.

* Explicitly implement `Hash` and `Eq`.
`protobuf.rs` is used for tests and is the expected output for prost macro.

- add tests

* suppress clippy warning, the logic is correct